+++ modules/triggers/TriggerCollection.cs
@@ -27,7 +27,9 @@ namespace AutoScreenCapture
         private const string TRIGGER_CONDITION = "condition";
         private const string TRIGGER_ACTION = "action";
         private const string TRIGGER_EDITOR = "editor";
-        private const string TRIGGER_XPATH = "/" + XML_FILE_ROOT_NODE + "/" + XML_FILE_TRIGGERS_NODE + "/" + XML_FILE_TRIGGER_NODE;
+
+        private const string TRIGGER_XPATH =
+            "/" + XML_FILE_ROOT_NODE + "/" + XML_FILE_TRIGGERS_NODE + "/" + XML_FILE_TRIGGER_NODE;
 
         public List<Trigger>.Enumerator GetEnumerator()
         {
@@ -36,7 +38,7 @@ namespace AutoScreenCapture
 
         IEnumerator IEnumerable.GetEnumerator()
         {
-            return ((IEnumerable<Trigger>)_triggerList).GetEnumerator();
+            return ((IEnumerable<Trigger>) _triggerList).GetEnumerator();
         }
 
         IEnumerator<Trigger> IEnumerable<Trigger>.GetEnumerator()
@@ -94,7 +96,8 @@ namespace AutoScreenCapture
         /// </summary>
         public void Load()
         {
-            if (File.Exists(FileSystem.ApplicationFolder + FileSystem.TriggersFile))
+            if (Directory.Exists(FileSystem.ApplicationFolder) &&
+                File.Exists(FileSystem.ApplicationFolder + FileSystem.TriggersFile))
             {
                 XmlDocument xDoc = new XmlDocument();
                 xDoc.Load(FileSystem.ApplicationFolder + FileSystem.TriggersFile);
@@ -119,12 +122,14 @@ namespace AutoScreenCapture
 
                                 case TRIGGER_CONDITION:
                                     xReader.Read();
-                                    trigger.ConditionType = (TriggerConditionType)Enum.Parse(typeof(TriggerConditionType), xReader.Value);
+                                    trigger.ConditionType =
+                                        (TriggerConditionType) Enum.Parse(typeof(TriggerConditionType), xReader.Value);
                                     break;
 
                                 case TRIGGER_ACTION:
                                     xReader.Read();
-                                    trigger.ActionType = (TriggerActionType)Enum.Parse(typeof(TriggerActionType), xReader.Value);
+                                    trigger.ActionType =
+                                        (TriggerActionType) Enum.Parse(typeof(TriggerActionType), xReader.Value);
                                     break;
 
                                 case TRIGGER_EDITOR:
@@ -146,11 +151,16 @@ namespace AutoScreenCapture
             else
             {
                 // Setup a few "built in" triggers by default.
-                Add(new Trigger("Application Startup -> Show", TriggerConditionType.ApplicationStartup, TriggerActionType.ShowInterface, string.Empty));
-                Add(new Trigger("Capture Started -> Hide", TriggerConditionType.ScreenCaptureStarted, TriggerActionType.HideInterface, string.Empty));
-                Add(new Trigger("Capture Stopped -> Show", TriggerConditionType.ScreenCaptureStopped, TriggerActionType.ShowInterface, string.Empty));
-                Add(new Trigger("Interface Closing -> Exit", TriggerConditionType.InterfaceClosing, TriggerActionType.ExitApplication, string.Empty));
-                Add(new Trigger("Limit Reached -> Stop", TriggerConditionType.LimitReached, TriggerActionType.StopScreenCapture, string.Empty));
+                Add(new Trigger("Application Startup -> Show", TriggerConditionType.ApplicationStartup,
+                    TriggerActionType.ShowInterface, string.Empty));
+                Add(new Trigger("Capture Started -> Hide", TriggerConditionType.ScreenCaptureStarted,
+                    TriggerActionType.HideInterface, string.Empty));
+                Add(new Trigger("Capture Stopped -> Show", TriggerConditionType.ScreenCaptureStopped,
+                    TriggerActionType.ShowInterface, string.Empty));
+                Add(new Trigger("Interface Closing -> Exit", TriggerConditionType.InterfaceClosing,
+                    TriggerActionType.ExitApplication, string.Empty));
+                Add(new Trigger("Limit Reached -> Stop", TriggerConditionType.LimitReached,
+                    TriggerActionType.StopScreenCapture, string.Empty));
 
                 Save();
             }
@@ -161,48 +171,52 @@ namespace AutoScreenCapture
         /// </summary>
         public void Save()
         {
-            XmlWriterSettings xSettings = new XmlWriterSettings();
-            xSettings.Indent = true;
-            xSettings.CloseOutput = true;
-            xSettings.CheckCharacters = true;
-            xSettings.Encoding = Encoding.UTF8;
-            xSettings.NewLineChars = Environment.NewLine;
-            xSettings.IndentChars = XML_FILE_INDENT_CHARS;
-            xSettings.NewLineHandling = NewLineHandling.Entitize;
-            xSettings.ConformanceLevel = ConformanceLevel.Document;
-
-            if (File.Exists(FileSystem.ApplicationFolder + FileSystem.TriggersFile))
+            if (Directory.Exists(FileSystem.ApplicationFolder))
             {
-                File.Delete(FileSystem.ApplicationFolder + FileSystem.TriggersFile);
-            }
+                XmlWriterSettings xSettings = new XmlWriterSettings();
+                xSettings.Indent = true;
+                xSettings.CloseOutput = true;
+                xSettings.CheckCharacters = true;
+                xSettings.Encoding = Encoding.UTF8;
+                xSettings.NewLineChars = Environment.NewLine;
+                xSettings.IndentChars = XML_FILE_INDENT_CHARS;
+                xSettings.NewLineHandling = NewLineHandling.Entitize;
+                xSettings.ConformanceLevel = ConformanceLevel.Document;
 
-            using (XmlWriter xWriter = XmlWriter.Create(FileSystem.ApplicationFolder + FileSystem.TriggersFile, xSettings))
-            {
-                xWriter.WriteStartDocument();
-                xWriter.WriteStartElement(XML_FILE_ROOT_NODE);
-                xWriter.WriteAttributeString("app", "version", XML_FILE_ROOT_NODE, Settings.ApplicationVersion);
-                xWriter.WriteAttributeString("app", "codename", XML_FILE_ROOT_NODE, Settings.ApplicationCodename);
-                xWriter.WriteStartElement(XML_FILE_TRIGGERS_NODE);
-
-                foreach (object obj in _triggerList)
+                if (File.Exists(FileSystem.ApplicationFolder + FileSystem.TriggersFile))
                 {
-                    Trigger trigger = (Trigger)obj;
-
-                    xWriter.WriteStartElement(XML_FILE_TRIGGER_NODE);
-                    xWriter.WriteElementString(TRIGGER_NAME, trigger.Name);
-                    xWriter.WriteElementString(TRIGGER_CONDITION, trigger.ConditionType.ToString());
-                    xWriter.WriteElementString(TRIGGER_ACTION, trigger.ActionType.ToString());
-                    xWriter.WriteElementString(TRIGGER_EDITOR, trigger.Editor);
-
-                    xWriter.WriteEndElement();
+                    File.Delete(FileSystem.ApplicationFolder + FileSystem.TriggersFile);
                 }
 
-                xWriter.WriteEndElement();
-                xWriter.WriteEndElement();
-                xWriter.WriteEndDocument();
+                using (XmlWriter xWriter =
+                    XmlWriter.Create(FileSystem.ApplicationFolder + FileSystem.TriggersFile, xSettings))
+                {
+                    xWriter.WriteStartDocument();
+                    xWriter.WriteStartElement(XML_FILE_ROOT_NODE);
+                    xWriter.WriteAttributeString("app", "version", XML_FILE_ROOT_NODE, Settings.ApplicationVersion);
+                    xWriter.WriteAttributeString("app", "codename", XML_FILE_ROOT_NODE, Settings.ApplicationCodename);
+                    xWriter.WriteStartElement(XML_FILE_TRIGGERS_NODE);
 
-                xWriter.Flush();
-                xWriter.Close();
+                    foreach (object obj in _triggerList)
+                    {
+                        Trigger trigger = (Trigger) obj;
+
+                        xWriter.WriteStartElement(XML_FILE_TRIGGER_NODE);
+                        xWriter.WriteElementString(TRIGGER_NAME, trigger.Name);
+                        xWriter.WriteElementString(TRIGGER_CONDITION, trigger.ConditionType.ToString());
+                        xWriter.WriteElementString(TRIGGER_ACTION, trigger.ActionType.ToString());
+                        xWriter.WriteElementString(TRIGGER_EDITOR, trigger.Editor);
+
+                        xWriter.WriteEndElement();
+                    }
+
+                    xWriter.WriteEndElement();
+                    xWriter.WriteEndElement();
+                    xWriter.WriteEndDocument();
+
+                    xWriter.Flush();
+                    xWriter.Close();
+                }
             }
         }
     }
